refactored WeightedStaticSchedulingWithSetups
